requirements.txt
Flask==2.0.3
Pillow==8.4.0
pillow-avif-plugin==1.4.6
werkzeug==2.2.2